remove task endpoint

+++ server/routes/taskRouter.js
@@ -6,29 +6,14 @@ const router = express.Router()
 
 
 // Remove Task
-router.delete("/task/:id", async (req, res) => {
+router.delete("/task/:id", auth, async (req, res) => {
     try {
-        const task = await Task.findOneAndRemove({ _id: req.params.id })
+        const task = await Task.findOneAndRemove({ _id: req.params.id, owner: req.user._id })
         if (!task) throw new Error()
         res.send(task).status(200)
     } catch (error) {
         res.status(500).send(error)
     }
-})
-
-// Read Task
-router.get("/task", async (req, res) => {
-    const task = await Task.find({})
-
-    try {
-
-        res.send(task).status(200)
-
-    } catch (error) {
-
-        res.status(500).send("Didn't find any tasks", error)
-    }
-
 })
 
 
@@ -66,17 +51,13 @@ router.get("/task/me", auth, async (req, res) => {
     const user = await req.user
 
     try {
-
         await user.populate("tasks")
-
         // const userTasks = await req.user.tasks
         res.send(user.tasks).status(200)
 
     } catch (error) {
         res.status(404).send(error)
     }
-
-
 })
 
 
@@ -89,8 +70,6 @@ router.post("/task", auth, async (req, res) => {
     })
 
     try {
-
-
         await task.save()
         res.status(201).send(task)
